# core/docs/swagger.yaml
definitions:
...
  controller.AddTokenRequest:
    properties:
      expiredAt:
        type: integer
      models:
        items:
          type: string
        type: array
      name:
        type: string
      quota:
        type: number
      subnets:
        items:
          type: string
        type: array
    type: object
...
  controller.TokenResponse:
    properties:
      accessed_at:
        type: string
      created_at:
        type: string
      expired_at:
        type: string
      group:
        type: string
      id:
        type: integer
      key:
        type: string
      models:
        items:
          type: string
        type: array
      name:
        type: string
      quota:
        type: number
      request_count:
        type: integer
      status:
        type: integer
      subnets:
        items:
          type: string
        type: array
      used_amount:
        type: number
    type: object
...
  controller.UpdateTokenNameRequest:
    properties:
      name:
        type: string
    type: object
  controller.UpdateTokenStatusRequest:
    properties:
      status:
        type: integer
    type: object
...
  middleware.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  title: AI Proxy Swagger API
  version: "1.0"
paths:
...
  /api/token/{group}:
    post:
      consumes:
      - application/json
      description: Adds a new token to a specific group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Auto create group
        in: query
        name: auto_create_group
        type: boolean
      - description: Ignore exist
        in: query
        name: ignore_exist
        type: boolean
      - description: Token information
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/controller.AddTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.TokenResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Add group token
      tags:
      - token
  /api/token/{group}/{id}:
    delete:
      description: Deletes a specific token from a group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete group token
      tags:
      - token
    get:
      description: Returns detailed information about a specific token for a specific
        group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.TokenResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get token by ID for a specific group
      tags:
      - token
    put:
      consumes:
      - application/json
      description: Updates an existing token in a specific group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated token information
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/controller.AddTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.TokenResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update group token
      tags:
      - token
  /api/token/{group}/{id}/name:
    post:
      consumes:
      - application/json
      description: Updates the name of a token in a specific group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      - description: Name information
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateTokenNameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group token name
      tags:
      - token
  /api/token/{group}/{id}/status:
    post:
      consumes:
      - application/json
      description: Updates the status of a token in a specific group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status information
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateTokenStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update group token status
      tags:
      - token
  /api/token/{group}/batch_delete:
    post:
      description: Deletes multiple tokens from a specific group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Token IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete group tokens
      tags:
      - token
  /api/token/{group}/search:
    get:
      description: Returns a paginated list of tokens for a specific group based on
        search criteria
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Keyword
        in: query
        name: keyword
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Order
        in: query
        name: order
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: Key
        in: query
        name: key
        type: string
      - description: Status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        tokens:
                          items:
                            $ref: '#/definitions/controller.TokenResponse'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Search tokens for a specific group
      tags:
      - token
  /api/tokens/:
    get:
      description: Returns a paginated list of all tokens
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Group name
        in: query
        name: group
        type: string
      - description: Order
        in: query
        name: order
        type: string
      - description: Status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        tokens:
                          items:
                            $ref: '#/definitions/controller.TokenResponse'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all tokens
      tags:
      - tokens
  /api/tokens/{group}:
    get:
      description: Returns a paginated list of all tokens for a specific group
      parameters:
      - description: Group name
        in: path
        name: group
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Order
        in: query
        name: order
        type: string
      - description: Status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        tokens:
                          items:
                            $ref: '#/definitions/controller.TokenResponse'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all tokens for a specific group
      tags:
      - tokens
  /api/tokens/{id}:
    delete:
      description: Deletes a specific token by ID
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete token
      tags:
      - tokens
    get:
      description: Returns detailed information about a specific token
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.TokenResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get token by ID
      tags:
      - tokens
    put:
      consumes:
      - application/json
      description: Updates an existing token's information
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated token information
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/controller.AddTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/controller.TokenResponse'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update token
      tags:
      - tokens
  /api/tokens/{id}/name:
    post:
      consumes:
      - application/json
      description: Updates the name of a specific token
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      - description: Name information
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateTokenNameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update token name
      tags:
      - tokens
  /api/tokens/{id}/status:
    post:
      consumes:
      - application/json
      description: Updates the status of a specific token
      parameters:
      - description: Token ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status information
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateTokenStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update token status
      tags:
      - tokens
  /api/tokens/batch_delete:
    post:
      consumes:
      - application/json
      description: Deletes multiple tokens by their IDs
      parameters:
      - description: Token IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete multiple tokens
      tags:
      - tokens
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"